COMP: Remove warning

+++ VTK/Common/vtkMath.cxx
@@ -2911,7 +2911,7 @@ void vtkMath::SpiralPoints(vtkIdType num, vtkPoints * offsets)
 }
 
 //----------------------------------------------------------------------------
-int vtkMath::Solve3PointCircle(const double p1[3], const double p2[3], 
+double vtkMath::Solve3PointCircle(const double p1[3], const double p2[3], 
                                const double p3[3], double center[3])
 {
   double v21[3], v32[3], v13[3];
